Broke option into separate component to deal with state issues on Select item. Still need to fix column width but basics for equipping a mod seem to all be in place.
